doc, tools: extract loadGtoc
